# merge GTEx-DESeq branch back to master
Project Status
1. fixed in SalmonPairedRead.wdl task
   - the paired fastq files where backwards

2. GTEx
   - re-computed all salmon quant file
   - re-create count matrix and colData. Files copied to $kl/GTEx on private

3. TCGA
   - most of re-compute has been done
   - problem work sets
     - DLBC
     - GBM
     - LAML
     - SARC
     - BRCA
     - strange data sets. most samples run, some fail. Need to explore failures
       * COAD 192 failed sample
       * READ 72 failed samples
       * STAD 39 failed samples
       * UCEC 380 failed samples

4. 1vsAll DESeq docker created
   - TODO
     *  test using cromwell & wdl
     * add lcfshrink
     * save coeficients? can we calculate correlation of dependicy for each signifigant gene
       + how much variation did the glm account for

5. CIBER Sort
   - TODO
     * evaluate tissue signatures.
       + should we drop some tissues
       + should we lump some tissues
     * use validation set to identify best top n value

+++ bin/gcpSpark/deleteCluster.sh
@@ -12,6 +12,13 @@ configFile=$2
 set -x
 
 source $configFile
+status=$?
+if [ $status -ne 0 ]
+then
+    echo "ERROR unable to source ${configFile}"
+    exit 1
+fi
+
 gcloud dataproc clusters delete $clusterName \
        --project=${PROJECT_ID} \
        --region=${REGION}
